dodanie wrappera w Adzie do biblioteki w C

+++ C/biblioteka/biblioteka.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "biblioteka.h"
 
-unsigned long long silnia_rekurencja(int n){
+int silnia_rekurencja(int n){
 	if (n == 0 || n == 1){
 		return 1;
 	} else {
@@ -9,15 +9,15 @@ unsigned long long silnia_rekurencja(int n){
 	}
 }
 
-unsigned long long silnia_petla(int n){
-	unsigned long long wynik = 1;
+int silnia_petla(int n){
+	int wynik = 1;
 	for (int i = 1; i <= n; i++){
 		wynik = wynik*i;
 	}
 	return wynik;
 }
 
-unsigned int NWD_rekurencja(int a, int b){
+int NWD_rekurencja(int a, int b){
 	if(b == 0){
 		return a;
 	} else {
@@ -25,7 +25,7 @@ unsigned int NWD_rekurencja(int a, int b){
 	}
 }
 
-unsigned int NWD_petla(int a, int b){
+int NWD_petla(int a, int b){
 	while(b != 0){
 		int temp = b;
 		b = a % b;
